Added all the necessary classes and changed the project to use floats instead of doubles as that much precision isn't needed
Former-commit-id: a91395c6f7f2f8cd99fe5b5b1dd541b3cb18d0ff

+++ project/Agents/qlearning.cpp
@@ -13,7 +13,7 @@ QLearning<C>::QLearning(C controller, float epsilon, float gamma):
 
     //Initialising the Q fonction to 0 for each state action pair
 
-    qvalues = vector<double>(controller.saPairSpaceSize(),0);
+    qvalues = vector<float>(controller.saPairSpaceSize(),0);
 }
 
 template <class C>
@@ -25,13 +25,13 @@ void QLearning<C>::initialiseEpisode()
 template <class C>
 void QLearning<C>::greedyPolicy()
 {
-    vector<double> possibleQValues;
+    vector<float> possibleQValues;
 
     for (int i=0;i<Agent<C>::actions().cardinal();i++)
     {
         possibleQValues.push_back(qvalues[this->controller.stateId(this->currentState())*this->actions().cardinal()+i]);
     }
-    double maxQValue = *max_element(possibleQValues.begin(),possibleQValues.end());
+    float maxQValue = *max_element(possibleQValues.begin(),possibleQValues.end());
 
     //If several qvalues are equal to the max, pick one of them randomly
 
@@ -46,7 +46,7 @@ void QLearning<C>::greedyPolicy()
     default_random_engine generator(random_device{}());
     uniform_int_distribution<int> dist(0,indexOfMax.size()-1);
     int actionId = indexOfMax[dist(generator)];
-    this->controller.setTakenAction(this->actions().actionFromId(actionId,new vector<double>()));
+    this->controller.setTakenAction(this->actions().actionFromId(actionId,new vector<float>()));
 }
 
 template <class C>
@@ -64,12 +64,12 @@ void QLearning<C>::updatePolicy()
     }
     else
     {
-        vector<double> updateChoice;
+        vector<float> updateChoice;
         for (int i=0;i<this->actions().cardinal();i++)
         {
             updateChoice.push_back(qvalues[csIndex+i]);
         }
-        double bestChoice = *max_element(updateChoice.begin(),updateChoice.end());
+        float bestChoice = *max_element(updateChoice.begin(),updateChoice.end());
         qvalues[psIndex+actionId] = (1./(this->episodeNumber+1))*(this->takenReward()+gamma*bestChoice-qvalues[psIndex+actionId]);
     }
 }
